chore(move): colors to utils
